pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "repolish"
version = "0.1.0"
description = "GitHub README generator using CrewAI"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "crewai[tools]>=0.130.0",
    "PyGithub>=1.59.0",
    "gradio>=4.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0.2",
    "markdown>=3.8.2",
]
requires-python = ">=3.10,<3.14"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
run_crew = "repolish.main:run"

[tool.uv]
dev-dependencies = []
